requirements.txt
numpy==1.22.3
tqdm==4.64.0
pyyaml==5.4.1
tensorboardx==2.5.1
neptune-client==0.16.7
neptune-sacred==0.10.0
sacred==0.8.2
webdataset==0.1.103
scikit-image==0.19.3
timm==0.4.12
fairscale==0.4.4
einops==0.3.0
ffmpeg-python==0.2.0
matplotlib==3.6.0
simplejson==3.18
pycocoevalcap
pycocotools


accelerate==0.18.0
sentencepiece==0.1.97
chardet==5.1.0
git+https://github.com/huggingface/transformers@v4.29.0
fschat==0.1.10